Formatting

+++ genericskill/SkillBindingSignet.java
@@ -3,8 +3,7 @@ package genericskill;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.ChunkCoordinates;
 
-public class SkillBindingSignet extends SkillGeneric{
-
+public class SkillBindingSignet extends SkillGeneric {
 	@Override
 	public String getDescription() {
 		return "Teleports you back to your\nroots.\n(requires a Heritage Amulet)";
@@ -32,21 +31,21 @@ public class SkillBindingSignet extends SkillGeneric{
 
 	@Override
 	public boolean onSkillTrigger(EntityPlayer player) {
-		if(!canPlayerUseSkill(player))
+		if (!canPlayerUseSkill(player))
 			return false;
-    	if(!player.worldObj.isRemote){
-    		if(!player.worldObj.provider.canRespawnHere())
-    			player.travelToDimension(0);
-	    	ChunkCoordinates chunkcoordinates = player.worldObj.getSpawnPoint();
-	    	int posX=chunkcoordinates.posX;
-	    	int posY=chunkcoordinates.posY;
-	    	int posZ=chunkcoordinates.posZ;
-	    	while(!player.worldObj.isAirBlock(posX, posY, posZ))
-	    		posY++; //So you don't spawn in the floor.
-			player.setPositionAndUpdate((float)posX+0.5F, (float)posY+0.1F, (float)posZ+0.5F);
-    	}
+		if (!player.worldObj.isRemote) {
+			if (!player.worldObj.provider.canRespawnHere())
+				player.travelToDimension(0);
+			ChunkCoordinates chunkcoordinates = player.worldObj.getSpawnPoint();
+			int posX = chunkcoordinates.posX;
+			int posY = chunkcoordinates.posY;
+			int posZ = chunkcoordinates.posZ;
+			while (!player.worldObj.isAirBlock(posX, posY, posZ))
+				posY++; //So you don't spawn in the floor.
+			player.setPositionAndUpdate(posX + 0.5F, posY + 0.1F, posZ + 0.5F);
+		}
 		player.inventory.consumeInventoryItem(GenericSkills.heritageAmulet.itemID);
-    	player.worldObj.playSoundAtEntity(player, "mob.ghast.fireball", 1.0F, 1.0F);
+		player.worldObj.playSoundAtEntity(player, "mob.ghast.fireball", 1.0F, 1.0F);
 		return true;
 	}
 }
